+++ src/app/screens/Dashboard/index.js
@@ -2,15 +2,19 @@ import React from 'react';
 import { Switch } from 'react-router-dom';
 
 import RouteItem from '~components/Routes/components/RouteItem';
+import Navbar from '~components/Navbar';
 import PATHS from '~components/Routes/paths';
 
 import Home from './screens/Home';
 
 function Dashboard() {
   return (
-    <Switch>
-      <RouteItem exact path={PATHS.home} component={Home} />
-    </Switch>
+    <>
+      <Navbar />
+      <Switch>
+        <RouteItem exact path={PATHS.home} component={Home} />
+      </Switch>
+    </>
   );
 }
